revert moving loadTheme related code
