+++ packages/styling/etc/styling.api.md
@@ -14,7 +14,6 @@ import { DefaultPalette } from '@fluentui/theme/lib/colors/DefaultPalette';
 import { fontFace } from '@uifabric/merge-styles';
 import { FontSizes } from '@fluentui/theme/lib/fonts';
 import { FontWeights } from '@fluentui/theme/lib/fonts';
-import { getTheme } from '@fluentui/theme';
 import { IconFontSizes } from '@fluentui/theme/lib/fonts';
 import { ICSPSettings } from '@uifabric/merge-styles';
 import { ICustomizerContext } from '@uifabric/utilities';
@@ -37,14 +36,10 @@ import { IStyleSet } from '@uifabric/merge-styles';
 import { IStyleSheetConfig } from '@uifabric/merge-styles';
 import { ITheme } from '@fluentui/theme/lib/types/ITheme';
 import { keyframes } from '@uifabric/merge-styles';
-import { loadTheme } from '@fluentui/theme';
 import { mergeStyles } from '@uifabric/merge-styles';
 import { mergeStyleSets } from '@uifabric/merge-styles';
 import { registerDefaultFontFaces } from '@fluentui/theme/lib/fonts/DefaultFontStyles';
-import { registerOnThemeChangeCallback } from '@fluentui/theme';
-import { removeOnThemeChangeCallback } from '@fluentui/theme';
 import { Stylesheet } from '@uifabric/merge-styles';
-import { ThemeSettingName } from '@fluentui/theme';
 
 // @public (undocumented)
 export const AnimationClassNames: {
@@ -133,7 +128,8 @@ export function getPlaceholderStyles(styles: IStyle): IStyle;
 // @public (undocumented)
 export function getScreenSelector(min: number, max: number): string;
 
-export { getTheme }
+// @public
+export function getTheme(depComments?: boolean): ITheme;
 
 // Warning: (ae-internal-missing-underscore) The name "getThemedContext" should be prefixed with an underscore because the declaration is marked as @internal
 //
@@ -339,7 +335,8 @@ export { ITheme }
 
 export { keyframes }
 
-export { loadTheme }
+// @public
+export function loadTheme(theme: IPartialTheme, depComments?: boolean): ITheme;
 
 export { mergeStyles }
 
@@ -366,9 +363,11 @@ export function registerIconAlias(iconName: string, mappedToName: string): void;
 // @public
 export function registerIcons(iconSubset: IIconSubset, options?: Partial<IIconOptions>): void;
 
-export { registerOnThemeChangeCallback }
+// @public
+export function registerOnThemeChangeCallback(callback: (theme: ITheme) => void): void;
 
-export { removeOnThemeChangeCallback }
+// @public
+export function removeOnThemeChangeCallback(callback: (theme: ITheme) => void): void;
 
 // @public (undocumented)
 export const ScreenWidthMaxLarge: number;
@@ -411,7 +410,8 @@ export function setIconOptions(options: Partial<IIconOptions>): void;
 
 export { Stylesheet }
 
-export { ThemeSettingName }
+// @public (undocumented)
+export const ThemeSettingName = "theme";
 
 // @public
 export function unregisterIcons(iconNames: string[]): void;
