+++ packages/theme/etc/theme.api.md
@@ -158,9 +158,6 @@ export namespace FontWeights {
     bold: IFontWeight;
 }
 
-// @public
-export function getTheme(depComments?: boolean): ITheme;
-
 // @public (undocumented)
 export namespace IconFontSizes {
     const // (undocumented)
@@ -213,9 +210,6 @@ export interface IFontStyles {
     // @deprecated (undocumented)
     xxLargePlus: IRawStyle;
 }
-
-// @public (undocumented)
-export function initializeThemeInCustomizations(): void;
 
 // @public
 export interface IPalette {
@@ -452,9 +446,6 @@ export interface ITheme extends IScheme {
     };
 }
 
-// @public
-export function loadTheme(partialTheme: IPartialTheme, depComments?: boolean): ITheme;
-
 // @public (undocumented)
 export namespace LocalizedFontFamilies {
     const // (undocumented)
@@ -642,12 +633,6 @@ export type RecursivePartial<T> = {
 // @public (undocumented)
 export function registerDefaultFontFaces(baseUrl: string): void;
 
-// @public
-export function registerOnThemeChangeCallback(callback: (theme: ITheme) => void): void;
-
-// @public
-export function removeOnThemeChangeCallback(callback: (theme: ITheme) => void): void;
-
 // @public (undocumented)
 export namespace SharedColors {
     const // (undocumented)
@@ -732,9 +717,6 @@ export interface Theme extends ITheme {
     // (undocumented)
     tokens?: Tokens;
 }
-
-// @public (undocumented)
-export const ThemeSettingName = "theme";
 
 // @public (undocumented)
 export interface Tokens {
