+++ packages/theme/src/index.ts
@@ -6,4 +6,3 @@ export * from './spacing/index';
 export * from './motions/index';
 export * from './fonts/index';
 export * from './createTheme';
-export * from './globalTheme';
